backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.1.0
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.29.7
python-multipart==0.0.6
dj-database-url==2.1.0
# RAG dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
